Criação no contrato de serviço e criando put de forma pagamaneto com cartão de cŕedito

+++ src/main/java/com/example/findme/findme/domain/CartaoDeCredito.java
@@ -5,6 +5,7 @@ import lombok.*;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigInteger;
 
 @Getter
 @Setter
@@ -26,7 +27,7 @@ public class CartaoDeCredito {
 
     @NotNull
     @Column(name = "numero_do_cartao")
-    private Integer numeroCartao;
+    private BigInteger numeroCartao;
 
     @NotNull
     @Column(name = "validade_do_cartao")
